stable_diffusion/script_run_diffusion.py
import os
import argparse
import subprocess
import torch
import numpy as np
from tqdm import tqdm

def main():
    # Parse arguments
    parser = argparse.ArgumentParser()
    parser.add_argument('--num_images', type=int, default=10, help='Number of images to generate')
    parser.add_argument('--batch_size', type=int, default=1, help='Batch size')
    args = parser.parse_args()

    # Load the diffusion model
    model_path = '/stable_diffusion/modelstore/sd-v1-4.ckpt'
    diffusion = model.load(model_path)

    # Load the prompts
    prompts_path = '/input/prompts.txt'
    with open(prompts_path, 'r') as f:
        prompts = f.readlines()

    # Generate the images
    device = torch.device('cuda')
    batch_size = args.batch_size
    for i in tqdm(range(0, args.num_images, batch_size)):
        batch_prompts = prompts[i:i+batch_size]
        cmd = ['python', 'inference.py', '--diffusion_model', model_path, '--prompts'] + batch_prompts
        subprocess.run(cmd, check=True)
